Committed more WIP to repo (dangerous roads ahead)
I am developing this stuff on three different machines and
keeping the various tweaks and mods in place has proven to be a problem.

And a structure is beginning to emerge but I'm still a long way
from having anything comprehensive enough for anyone but me to use.

That said, the qfq stuff in wip is looking promising, as is the qfq-red
version, and the debloat script are not horrible either.