4 Mbit shaper plus fair queuing plus AQM that is fair to users
I have spent months in search of the holy grail here. I wanted a
home router shaper that was:

Fair to multiple users - if you had 2 users, they should get a max
of 1/2 the bandwidth, 3, 1/3 bandwidth, etc. In other words, it
needed to be bittorrent proof.

Did Fair queueing within streams from a user - So that a given
user with multiple forms of traffic had no head of line problem

Was additionally prioritizable for special traffic, like voice,
DNS from the router, etc.

Settable to something close to the observed upload bandwidth.

Did active queue management to handle big streams before they
got out of hand.

Ran on everything.

I came closest with this commit of htb-bt-qfq-red-4mbit. After
hammering down device buffers to 3000 with byte queue limits
and synthetically creating a 4mbit bottleneck with htb, I
layered on per ip QFQ fair queuing, and then fair queued within that
and then dumped the resulting queues into red...

it works really, really well.

The biggest thing that caused me to pull my hair out is that with a
single iperf stream competing with ping, ping would occilate wildly
in RTT from 2 to 80+ ms, when in a strictly fair universe it should have
stayed below 2ms. And thus I assumed I'd had bugs in the script
(which I did), which I kept hammering at... until there were no
bugs left and I kept getting the same lousy results.

Then I tried multiple streams, competing with ping... and ping times
stablized at ~10ms... even with 10 streams competing, ping stayed at
10ms... which happens to be the clock rate interrupt!

I had noticed this problem with htb and friends years ago, and one of
the first things I tried with cerowrt was to bump the clock rate up,
but failed as a given device driver's recognition routine failed at
1000HZ.

This testing above was done on a laptop with dual cores and bql-3
installed, as bql is not yet working on openwrt. bql might not
be strictly necessary, actually.

As to why it occilates when only with a single stream? Good question.
My assumption is that there is an 'optimization' within linux that
bypasses the qdisc in situations of low traffic.

More testing is indicated in more complex environments.

Also, as there are many hard coded constants in here, it will not scale
up well, nor down. There is presently no download speed limiter in
place either, this is complex enough as it stands, and I will be
synthesizing more models as time goes by.