Pure QFQ AQM solution for wireless and 100Mbit ethernet
The staqfq.lua code can now be renamed and made standard
in /etc/network/if-up-* on a normal desktop or server box
running a bql enabled kernel and git head of iproute2.

In addition to being the first version of this code to
actually work well (reducing interstream latencies by
two orders of magnitude on 100Mbit ethernet on a bql
enabled box)

This is the first version of this code re-written in lua.
Lua's advantages:

* More powerful than either C or shell (closures! yea!)
* Std language on openwrt, easily used on other systems
* Tiny language in either case

The code is now far more easy to read than shell and nearly
as fast as C. It also now uses tc's bulk load facility, and
loads the rules in under a second.

Also included in this commit is a tool to make reading tc -s qdisc
output mildly easier.