Slammed a bunch of adhoc tests I run into the repository
As I've finally settled on a set of useful benchmarks for measuring
latency under load, it's time to clean them up, and make them
repeatable for others.

That does not mean that that's what is in this commit.

I'm not proud of committing crappy code, but when you are crashing
your main development box on an hourly basis, it is a rather
good idea to NOT keep anything in just one place.