Modularized lua code and transliterated old diffserv code from sh
Created a default 'cero' module to toss various abstractions into
(for now). Will refactor as I get more familiar with the openwrt
environment.

Rewrote the old diffserv code (which was in the ash shell) into
lua. There is great potential here, not quite realized and
not quite debugged, yet.

There is a diffserv 'module' for the basic namespace, and
then a 'classify' program to generate the rules.

modularized the 'network renumber' script as well.

I have a huge backlog of little utilties to slam into this