Basic Latency under load results - 100Mbit, pfifo_fast, sfq, qfq
I finally assembled enough infrastructure to do a test, collect
the data, and plot the results in a form that is easily understood.

Driven by org-mode. I still would like to generate the plots
statically and link them into the text and do html...

and there are zillion other tests to run, but the highlights
are:

WIRED
TXQUEUELEN=1000
BANDWIDTH=100Mbit
BQL=4500 (byte queue limits)

1 iPerf induces 10-12ms latency w/pfifo_fast
10 iPerfs induces ~107 ms latency w/pfifo fast
10 IPerfs induces ~2.8 ms latency w/sfq
10 iPerfs induces ~1.7 ms latency with staqfq