Wireless QFQ tests.
The data is REALLY noisy and I need to run for a much longer time,
and come up with a good way to plot the results.

That said, QFQ has an average ping time of about 70ms vs 110 for pfifo_fast.

I suspect that using two (thoroughly debloated) ath9k cards will work
better - the laptop has an iwl which is known to be problematic.

And I need to vary some more variables.