Reorg of subdirs,etc.
I need to come up with a cleaner way of doing these tests,
formally design each module, and automate.

Moved all the bql-14 related data into one place, where
it can get sorted through later. As the data was only
valid for 2 days before the code it was testing changed,
I may not bother, however, as an example of the different
sorts of tests and comparisons needed to be run, it should
be a useful thing to automate from.