An attempt to simulate voip a little bit.
I sat down and wondered what would happen with a 10ms stream
beating against a network with sfq and qfq and pfifo_fast on it.

Everything downstream (3 hops) is running old sfq, and I'm
running 'new' sfq with the head of line fix in it against
all that. The results are basically that the downstream
boxes do a pretty good job on streams regardless of the
fact they were running old sfq, but I still need to
check a few things and collect all the related data.

I just consider this a start towards coming up with
automated tests that can use cdf in order to show valid
results. The data is too noisy to look at otherwise,
and that also means that all parameters and results
need to be looked at...

I do not in any way draw any conclusions from these
plots.

For example, along the way, iperf started acting up for some
reason - it would terminate early... and I switched to
netperf because of that...