Figuring out the interrelationships of the new variables is hard.
In the original QFQ work, I started off with red (before red
was working right, and adaptive red wasn't implemented yet),
did a whole bunch of tests, determined red didn't look right,
mentioned it to eric, who then fixed red and went on tear.

In between, I'd fallen back to using pfifo with a limit in the
range of 12-24 for ethernet. 24 seemed to be a pretty good value...

Now we have ared, and ared in sfq, and I'm really at a loss as
to how to calculate the right values for this newfangled
combination of tools.

A value of 12 is not the right value. But the last series of tests
was all run with that, and so, I'm preserving that for the record.