Added 'kleinrock' function
I started this whole 800+ line exercise because I wanted to do
a sqrt in shell.

This begs the question of what's the delay, what's the bandwidth,
and how many flows is good.

It also begs the question of what units to keep things in, and
how to make sure that is consistent throughout. Decimal megabits
vs mebibits is a 5% difference and that 5% can ruin your whole day.

So I will use decimal kilobits throughout and make sure to convert
to bytes whenever needed.