Normalized the tc qdisc output to have a more regular syntax
It's very hard to read the output and understand if it is correct
or not, as it is, with tc, and I've gone to enormous lengths,
creating shorthand for it in the script, and trying to
create output that can be looked over with the mark #1 eyeball
for correctness

It's annoying that I can't suppress stderr when I want to, in lua.
There has got to be a way.

Maybe I can use tc change? Don't know.