Stub for openwrt shaper emulator
My end game is to be able to drive this script and do tests
against various existing tools to be able to gather results.

The current openwrt shaper/aqm solution is a combination
of iptables rules and tc, generated by a combination of
awk and shell, using a combination of HFSC, SFQ, and RED
that I don't understand. It is also flawed in that
it doesn't do ipv6.

I've been meaning to at least, fix it for ipv6, for quite
some time. Now I have the infrastructure to do so, and
just have to warp my brain into translating awk and shell
in lua, again, as well as come up with clean ways to
drive iptables.