Alter eqfr wireless params, start at port of openwrt qos model
After thinking about the effect of SFQ on aggregation, I decided
that the closest thing to a 'win' would be to have SFQ use a quantum
of 3000 bytes, so we can have a ton of acks or 2 large packets
in an agg by default.

For all I know, this will work better than the existing non-sfq
based system...

I also translated a function from the openwrt qos lib into lua
that will hopefully let me strip out and create qos rules.

I've got a good grip on the flaws of the openwrt qos system now,
and my desire to preserve it, even for reference, is declining
rapidly...