Trying to find useful operating points...
This HTB+SFQRED model (and the variants)

IFACE=eth0 QMODEL=fourtier INGRESS=1 UPLINK=400 DOWNLINK=4000 ./debloat

Hold bandwidth nicely down to 400k for an upload, and shares
fairly.

However, simultaneous ping exhibits up to 30ms worth of jitter.
It's not SFQ that's the problem....

It's my hope that the problem is merely that a higher clock
and hires timers, etc, will alleviate the issue...